pyproject.toml
[project]
name = "agent-force"
version = "0.1.0"
description = "Automated safety evaluation framework for AI agents"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
    "mcp>=1.0.0",
]

[project.optional-dependencies]
server = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "aiosqlite>=0.20.0",
    "httpx>=0.27.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
]

[project.scripts]
agent-force = "safety_kit.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["safety_kit", "agents", "sandbox_env"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["F401"]
